Pass all the string pieces to the host
Althought the proposed CSP behavior for `new Functoin` is
to only support the 1-parameter arg, there is no good reason
for us to explicitly enforce that behavior. Instead, there is a
desire to simply pass all the strings.

In a possible future extension with Trusted Types support,
we will need to pass all the non stringified pieces too.